# remove min/maxes from granule metadata