Since changing the way we compact granules in https://github.com/polarsignals/frostdb/pull/189 it's less obvious that granule level filtering is valuable due to the reduction in the number of granules.

Granule level filtering also added locking and has a current outstanding bug https://github.com/polarsignals/frostdb/issues/188

I propose that we remove the granule level filtering, and can add it back in the future should a need arise.

```
thor@thors-MacBook-Pro frostdb % benchstat before.txt after.txt
name                                          old time/op    new time/op    delta
QueryTypes-10                                   5.16ms ± 0%    4.88ms ± 0%   ~     (p=1.000 n=1+1)
QueryMerge-10                                   22.2ms ± 0%    21.7ms ± 0%   ~     (p=1.000 n=1+1)
QueryRange-10                                   12.0ms ± 0%    11.8ms ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_1000Rows_10Iters_10Writers-10      198ms ± 0%     134ms ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_1000Iters_1Writers-10      362ms ± 0%     337ms ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_100Iters_100Writers-10     34.9s ± 0%     35.0s ± 0%   ~     (p=1.000 n=1+1)

name                                          old alloc/op   new alloc/op   delta
QueryTypes-10                                   5.61MB ± 0%    5.61MB ± 0%   ~     (p=1.000 n=1+1)
QueryMerge-10                                   66.7MB ± 0%    66.4MB ± 0%   ~     (p=1.000 n=1+1)
QueryRange-10                                   27.9MB ± 0%    27.6MB ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_1000Rows_10Iters_10Writers-10      191MB ± 0%     187MB ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_1000Iters_1Writers-10     1.56GB ± 0%    1.56GB ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_100Iters_100Writers-10    15.8GB ± 0%    15.8GB ± 0%   ~     (p=1.000 n=1+1)

name                                          old allocs/op  new allocs/op  delta
QueryTypes-10                                    85.9k ± 0%     85.9k ± 0%   ~     (all equal)
QueryMerge-10                                     157k ± 0%      154k ± 0%   ~     (p=1.000 n=1+1)
QueryRange-10                                    57.7k ± 0%     54.1k ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_1000Rows_10Iters_10Writers-10       374k ± 0%      369k ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_1000Iters_1Writers-10      1.93M ± 0%     1.90M ± 0%   ~     (p=1.000 n=1+1)
_Table_Insert_100Rows_100Iters_100Writers-10     19.4M ± 0%     19.1M ± 0%   ~     (p=1.000 n=1+1)
```